requirements.txt
Flask
gunicorn
requests
msal
pdfplumber
pytesseract
pillow
opencv-python-headless
numpy
certifi
python-dotenv
pytesseract
opencv-python-headless
pillow
pandas
openpyxl
numpy
pdf2image
gspread
google-auth
pdfplumber
google-api-python-client